pyproject.toml
[project]
name = "py-pokemon-api"
version = "0.1.0"
description = ""
authors = [
    {name = "jorge.machado",email = "jorge.vmachado@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.11.0,<3.0.0)",
    "tzdata (>=2025.3,<2026.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.1,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "freezegun (>=1.5.5,<2.0.0)",
    "testcontainers (>=4.14.1,<5.0.0)"
]
[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'

[tool.coverage.run]
core = "ctrace"
concurrency = ["thread", "greenlet"]